feat: remove unnecessary tracked channel abstraction
